Commit 2
Added/updated makefile
Linked everything to new libft.h
updated memmove and part 2 code

+++ src/ft_memchr.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-void	*ft_memchr(void *buf, int c, unsigned int len)
+void	*ft_memchr(void *buf, int c, size_t len)
 {
 	unsigned char	*mem;
