+++ src/ft_memcmp.c
@@ -1,4 +1,6 @@
-int	ft_memcmp(const void *s1, const void *s2, unsigned int len)
+#include "Libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t len)
 {
 	const unsigned char	*str1;
 	const unsigned char	*str2;
