+++ src/ft_memset.c
@@ -1,4 +1,6 @@
-void	*ft_memset(void *dst, int c, unsigned int len)
+#include "Libft.h"
+
+void	*ft_memset(void *dst, int c, size_t len)
 {
 	unsigned char	*dest;
