+++ src/ft_strncmp.c
@@ -1,6 +1,8 @@
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+#include "Libft.h"
+
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	i;
+	int	i;
 
 	i = 0;
 	if (n == 0)
